# shared/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"Town": {
				"description": "The public-facing model that represents a town. More information about the town\nis available for users who join it.",
				"properties": {
					"friendlyName": {
						"type": "string",
						"description": "The name that users see on the landing page to determine which town to join"
					},
					"townID": {
						"type": "string",
						"description": "An internal ID that is used to uniquely identify each town"
					},
					"currentOccupancy": {
						"type": "number",
						"format": "double",
						"description": "The current number of players in this town"
					},
					"maximumOccupancy": {
						"type": "number",
						"format": "double",
						"description": "The maximum number of players allowed in this town"
					}
				},
				"required": [
					"friendlyName",
					"townID",
					"currentOccupancy",
					"maximumOccupancy"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InvalidParametersError": {
				"properties": {
					"code": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TownSettingsUpdate": {
				"properties": {
					"isPubliclyListed": {
						"type": "boolean"
					},
					"friendlyName": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "covey-town-townService",
		"version": "1.0.0",
		"description": "Backend service for processing Covey.Town town requests",
		"license": {
			"name": "BSD-3-Clause"
		},
		"contact": {
			"name": "Jonathan Bell"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/towns": {
			"get": {
				"operationId": "ListTowns",
				"responses": {
					"200": {
						"description": "list of towns",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Town"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all towns that are set to be publicly available",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTown",
				"responses": {
					"200": {
						"description": "The ID of the newly created town, and a secret password that will be needed to update or delete this town.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TownCreateResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"townID": "stringID",
											"townUpdatePassword": "secretPassword"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The public-facing information for the new town",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TownCreateParams",
								"description": "The public-facing information for the new town"
							},
							"example": {
								"friendlyName": "My testing town public name",
								"isPubliclyListed": true
							}
						}
					}
				}
			}
		},
		"/towns/{townID}": {
			"patch": {
				"operationId": "UpdateTown",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid password or update values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Updates an existing town's settings by ID",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "town to update",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "town update password, must match the password returned by createTown",
						"in": "header",
						"name": "X-CoveyTown-Password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The updated settings",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TownSettingsUpdate",
								"description": "The updated settings"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTown",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid password or update values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Deletes a town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town to delete",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "town update password, must match the password returned by createTown",
						"in": "header",
						"name": "X-CoveyTown-Password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/towns/{townID}/conversationArea": {
			"post": {
				"operationId": "CreateConversationArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a conversation area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new conversation area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must match the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new conversation area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConversationArea",
								"description": "The new conversation area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/viewingArea": {
			"post": {
				"operationId": "CreateViewingArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a viewing area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new viewing area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new viewing area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ViewingArea",
								"description": "The new viewing area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/posterSessionArea": {
			"post": {
				"operationId": "CreatePosterSessionArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a poster session area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new poster session area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new poster session area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PosterSessionArea",
								"description": "The new poster session area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/whiteboardArea/createCanvas": {
			"post": {
				"operationId": "CreateNewCanvas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Canvas"
								}
							}
						}
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a whiteboard area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new whiteboard area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new viewing area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Canvas",
								"description": "The new viewing area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/{posterSessionId}/imageContents": {
			"patch": {
				"operationId": "GetPosterAreaImageContents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Gets the image contents of a given poster session area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to get the poster session area image contents",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "interactable ID of the poster session",
						"in": "path",
						"name": "posterSessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/towns/{townID}/{posterSessionId}/incStars": {
			"patch": {
				"operationId": "IncrementPosterAreaStars",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Increment the stars of a given poster session area in a given town, as long as there is\na poster image. Returns the new number of stars.",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to get the poster session area image contents",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "interactable ID of the poster session",
						"in": "path",
						"name": "posterSessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/towns/{townID}/whiteboardArea/destroyCanvas": {
			"patch": {
				"operationId": "DestroyCanvas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Canvas"
								}
							}
						}
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}